adds Serial mock with tests
Also switches over the output of error results in PPT100 to be b"" instead of None.  This is more in line with what the actual serial output will produce and makes it much easier to integrate with the Serial mock class.  The class doesn't have to deal with the possibility of different output types from get_response.